Move confirmation and information about status into service

+++ src/app/administration/events/events.service.ts
@@ -1,19 +1,22 @@
-import { HttpClient, HttpResourceRef, httpResource } from '@angular/common/http';
+import { HttpClient, HttpResourceRef, httpResource, HttpErrorResponse } from '@angular/common/http';
 import { inject, Injectable } from '@angular/core';
 import { Observable } from 'rxjs';
 import { tap } from 'rxjs/operators';
 import { Event, EventCapacitySummary, EventCreate } from './events.types';
 import { StorageKeys } from 'src/app/tokens/storage.tokens';
 import { StorageService } from 'src/app/services/storage.service';
+import { ToastrService } from 'ngx-toastr';
 
 @Injectable({
   providedIn: 'root'
 })
 export class EventService {
+  readonly #API_PATH = 'events';
+  
   readonly #httpClient = inject(HttpClient);
 
-  readonly #API_PATH = 'events';
   readonly #storageService = inject(StorageService);
+  readonly #toastr = inject(ToastrService);
 
   readonly events = httpResource<Event[]>(
     () => this.#endpoint('/'),
@@ -72,12 +75,36 @@ export class EventService {
   }
 
   public delete(
-    id: string
-  ): Observable<void> {
-    return this.#httpClient.delete<void>(
-      this.#endpoint(`/${id}/`)
+    event: Event,
+    shouldConfirm = true,
+  ): void {
+    if (shouldConfirm && !confirm(`Are you sure to delete event "${event.name}"?`)) {
+      return;
+    }
+    this.#httpClient.delete<void>(
+      this.#endpoint(`/${event.id}/`)
     ).pipe(
       tap(() => this.events.reload())
-    );
+    ).subscribe({
+      next: () => {
+        this.#toastr.info(
+          event.name,
+          'Event deleted',
+          {
+            progressBar: true
+          }
+        );
+      },
+      error: (err: HttpErrorResponse) => {
+        console.error(err);
+        this.#toastr.error(
+          `Error: ${err.message}`,
+          'Event wasn\'t deleted!',
+          {
+            progressBar: true,
+          }
+        );
+      }
+    });
   }
 }
